Moves geoserver-specific url routes to the geoserver app.
* acls, resolve_user and layer_batch_download have been moved to /gs/acls, gs/resolve_user, and gs/download.
* The old routes (using the /layer prefix) still work, but should be removed once legacy code is updated.
