+++ geonode/layers/urls.py
@@ -18,7 +18,8 @@
 #
 #########################################################################
 
-from django.conf.urls.defaults import patterns, url
+from django.conf import settings
+from django.conf.urls.defaults import include, patterns, url
 
 js_info_dict = {
     'packages': ('geonode.layers',),
@@ -28,10 +29,7 @@ urlpatterns = patterns(
     'geonode.layers.views',
     url(r'^$', 'layer_list', name='layer_browse'),
     url(r'^tag/(?P<slug>[-\w]+?)/$', 'layer_tag', name='layer_browse_tag'),
-    url(r'^acls/?$', 'layer_acls', name='layer_acls'),
-    url(r'^resolve_user/?$', 'resolve_user', name='layer_resolve_user'),
     url(r'^upload$', 'layer_upload', name='layer_upload'),
-    url(r'^download$', 'layer_batch_download', name='layer_batch_download'),
     url(r'^(?P<layername>[^/]*)$', 'layer_detail', name="layer_detail"),
     url(r'^(?P<layername>[^/]*)/metadata$', 'layer_metadata',
         name="layer_metadata"),
@@ -42,3 +40,17 @@ urlpatterns = patterns(
     #    name='batch_permssions'),
     #url(r'^api/batch_delete/?$', 'batch_delete', name='batch_delete'),
 )
+
+# -- Deprecated url routes for Geoserver authentication -- remove after GeoNode 2.1
+# -- Use /gs/acls, gs/resolve_user/, gs/download instead
+if 'geonode.geoserver' in settings.INSTALLED_APPS:
+    urlpatterns = patterns('geonode.geoserver.views',
+        url(r'^acls/?$', 'layer_acls', name='layer_acls_dep'),
+        url(r'^resolve_user/?$', 'resolve_user', name='layer_resolve_user_dep'),
+        url(r'^download$', 'layer_batch_download', name='layer_batch_download_dep'),
+    ) + urlpatterns
+
+
+
+
+
