+++ geonode/security/middleware.py
@@ -15,7 +15,9 @@ class LoginRequiredMiddleware(object):
         reverse('jscat'),
         reverse('lang'),
         reverse('layer_acls'),
+        reverse('layer_acls_dep'),
         reverse('layer_resolve_user'),
+        reverse('layer_resolve_user_dep'),
         '/account/(?!.*(?:signup))', # block unauthenticated users from creating new accounts.
         '/static/*',
     )
